* coffcode.h (coff_write_object_contents): Add ATTRIBUTE_UNUSED to 	silence warning. 	* coffgen.c (coff_print_symbol): Fix warnings. 	(coff_find_nearest_line): Likewise. 	* cofflink.c (_bfd_coff_link_input_bfd): Likewise. 	* coff-alpha.c (alpha_convert_external_reloc): Likewise. 	* format.c (bfd_check_format_matches): Likewise. 	(bfd_set_format): Likewise. 	* coff-ia64.c: Add missing prototypes. 	* elf64-alpha.c (struct alpha_elf_link_hash_entry): Make addend 	signed to silence warnings. 	(elf64_alpha_relocate_section): Likewise. 	(elf64_alpha_find_reloc_at_ofs): Fix warnings. 	(elf64_alpha_add_symbol_hook): Likewise. 	(elf64_alpha_final_link): Likewise. 	(elf64_alpha_relax_section): Remove redundant assign to info.gotent. 	(elf64_alpha_merge_gots): Add ATTRIBUTE_UNUSED to unused args. 	(elf64_alpha_size_got_sections): Likewise. 	* elfxx-ia64.c: Add missing prototypes. 	(elfNN_ia64_relocate_section): Fix warning. 	(elfNN_ia64_unwind_entry_compare): Make params const.

+++ bfd/coff-alpha.c
@@ -1334,7 +1334,7 @@ alpha_convert_external_reloc (output_bfd, info, input_bfd, ext_rel, h)
 	  break;
 	}
 
-      if (r_symndx == -1)
+      if (r_symndx == (unsigned long) -1)
 	abort ();
 
       /* Add the section VMA and the symbol value.  */
@@ -1347,7 +1347,7 @@ alpha_convert_external_reloc (output_bfd, info, input_bfd, ext_rel, h)
       /* Change the symndx value to the right one for
 	 the output BFD.  */
       r_symndx = h->indx;
-      if (r_symndx == -1)
+      if (r_symndx == (unsigned long) -1)
 	{
 	  /* Caller must give an error.  */
 	  r_symndx = 0;
